Add files via upload
Problem Statement

The objective is to analyze the flight 
booking dataset obtained from a 
platform which is used to book flight 
tickets. A thorough study of the data 
will aid in the discovery of valuable 
insights that will be of enormous 
value to passengers. Apply EDA, 
statistical methods and Machine 
learning algorithms in order to get 
meaningful information from it.